+++ src/auth/auth0-provider-with-navigate.js
